further trying to self contain sources

+++ source/mclib/utilities.cpp
@@ -41,11 +41,11 @@ void drawRect(const RECT& area, uint32_t color)
 		v[0].x										  = (float)clientArea.left;
 		v[0].y										  = (float)clientArea.top;
 		v[1].x										  = (float)clientArea.left;
-		v[1].y = (float)clientArea.bottom;
-		v[2].x = (float)clientArea.right;
-		v[2].y = (float)clientArea.bottom;
-		v[3].x = (float)clientArea.right;
-		v[3].y = (float)clientArea.top;
+		v[1].y										  = (float)clientArea.bottom;
+		v[2].x										  = (float)clientArea.right;
+		v[2].y										  = (float)clientArea.bottom;
+		v[3].x										  = (float)clientArea.right;
+		v[3].y										  = (float)clientArea.top;
 		gos_SetRenderState(gos_State_Texture, 0);
 		gos_SetRenderState(gos_State_ZWrite, 0);
 		gos_SetRenderState(gos_State_ZCompare, 0);
@@ -54,19 +54,16 @@ void drawRect(const RECT& area, uint32_t color)
 	}
 }
 
-void drawEmptyRect(const RECT& area, uint32_t leftTopBorderColor,
-	uint32_t rightBottomBorderColor)
+void drawEmptyRect(const RECT& area, uint32_t leftTopBorderColor, uint32_t rightBottomBorderColor)
 {
 	gos_VERTEX v[4];
 	memset(v, 0, sizeof(v));
 	v[0].rhw = v[1].rhw = v[2].rhw = v[3].rhw = 1.0;
 	RECT clientArea							  = area;
 	if ((clientArea.left < 0 && clientArea.right < 0) ||
-		(clientArea.right > Environment.screenWidth &&
-			clientArea.left > Environment.screenWidth) ||
+		(clientArea.right > Environment.screenWidth && clientArea.left > Environment.screenWidth) ||
 		(clientArea.top < 0 && clientArea.bottom < 0) ||
-		(clientArea.top > Environment.screenHeight &&
-			clientArea.bottom > Environment.screenHeight))
+		(clientArea.top > Environment.screenHeight && clientArea.bottom > Environment.screenHeight))
 		return;
 	if (clientArea.left < 0)
 		clientArea.left = 0;
@@ -148,8 +145,8 @@ void StaticInfo::showGUIWindow(bool bShow)
 
 bool StaticInfo::isInside(int32_t mouseX, int32_t mouseY)
 {
-	if ((location[0].x) <= mouseX && location[3].x >= mouseX &&
-		location[0].y <= mouseY && location[1].y >= mouseY)
+	if ((location[0].x) <= mouseX && location[3].x >= mouseX && location[0].y <= mouseY &&
+		location[1].y >= mouseY)
 		return true;
 	return false;
 }
@@ -172,10 +169,8 @@ void StaticInfo::getData(puint8_t buffer)
 		uint32_t localU  = location[0].u * textureData.Width;
 		uint32_t localV  = location[0].v * textureData.Width;
 		// Make sure we don't fall off of the texture!
-		if ((localU > textureData.Width) ||
-			((localU + uWidth) > textureData.Width) ||
-			(localV > textureData.Width) ||
-			((localV + vHeight) > textureData.Width))
+		if ((localU > textureData.Width) || ((localU + uWidth) > textureData.Width) ||
+			(localV > textureData.Width) || ((localV + vHeight) > textureData.Width))
 		{
 			memset(buffer, 0, sizeof(uint32_t) * 32 * 32);
 			gos_UnLockTexture(gosID);
@@ -184,8 +179,7 @@ void StaticInfo::getData(puint8_t buffer)
 		for (size_t y = 0; y < vHeight; y++)
 		{
 			uint32_t* textureMemory =
-				textureData.pTexture +
-				(localU + ((localV + y) * textureData.Width));
+				textureData.pTexture + (localU + ((localV + y) * textureData.Width));
 			for (size_t x = 0; x < uWidth; x++)
 			{
 				*bufMem = *textureMemory;
@@ -197,8 +191,8 @@ void StaticInfo::getData(puint8_t buffer)
 	}
 }
 
-void StaticInfo::init(FitIniFile& file, PSTR blockName, int32_t hiResOffsetX,
-	int32_t hiResOffsetY, uint32_t neverFlush)
+void StaticInfo::init(FitIniFile& file, PSTR blockName, int32_t hiResOffsetX, int32_t hiResOffsetY,
+	uint32_t neverFlush)
 {
 	memset(location, 0, sizeof(location));
 	char fileName[256];
@@ -224,8 +218,7 @@ void StaticInfo::init(FitIniFile& file, PSTR blockName, int32_t hiResOffsetX,
 		FullPathFileName fullPath;
 		_strlwr(fileName);
 		fullPath.init(artPath, fileName, ".tga");
-		int32_t ID = mcTextureManager->loadTexture(
-			fullPath, gos_Texture_Alpha, 0, 0, 0x2);
+		int32_t ID	 = mcTextureManager->loadTexture(fullPath, gos_Texture_Alpha, 0, 0, 0x2);
 		textureHandle  = ID;
 		uint32_t gosID = mcTextureManager->get_gosTextureHandle(ID);
 		TEXTUREPTR textureData;
@@ -247,38 +240,30 @@ void StaticInfo::init(FitIniFile& file, PSTR blockName, int32_t hiResOffsetX,
 		location[k].y	= y;
 		location[k].z	= 0.f;
 		location[k].rhw  = .5;
-		location[k].u =
-			(float)u / (float)textureWidth + (.1f / (float)textureWidth);
-		location[k].v =
-			(float)v / (float)textureWidth + (.1f / (float)textureWidth);
+		location[k].u	= (float)u / (float)textureWidth + (.1f / (float)textureWidth);
+		location[k].v	= (float)v / (float)textureWidth + (.1f / (float)textureWidth);
 	}
 	location[3].x = location[2].x = x + width;
 	location[2].y = location[1].y = y + height;
 	location[2].u				  = location[3].u =
-		((float)(u + uWidth)) / ((float)textureWidth) +
-		(.1f / (float)textureWidth);
+		((float)(u + uWidth)) / ((float)textureWidth) + (.1f / (float)textureWidth);
 	location[1].v = location[2].v =
-		((float)(v + vHeight)) / ((float)textureWidth) +
-		(.1f / (float)textureWidth);
+		((float)(v + vHeight)) / ((float)textureWidth) + (.1f / (float)textureWidth);
 	if (bRotated)
 	{
-		location[0].u =
-			(u + uWidth) / (float)textureWidth + (.1f / (float)textureWidth);
+		location[0].u = (u + uWidth) / (float)textureWidth + (.1f / (float)textureWidth);
 		;
 		location[1].u = u / (float)textureWidth + (.1f / (float)textureWidth);
 		;
 		location[2].u = u / (float)textureWidth + (.1f / (float)textureWidth);
-		location[3].u =
-			(u + uWidth) / (float)textureWidth + (.1f / (float)textureWidth);
+		location[3].u = (u + uWidth) / (float)textureWidth + (.1f / (float)textureWidth);
 		location[0].v = v / (float)textureWidth + (.1f / (float)textureWidth);
 		;
 		location[1].v = v / (float)textureWidth + (.1f / (float)textureWidth);
 		;
-		location[2].v =
-			(v + vHeight) / (float)textureWidth + (.1f / (float)textureWidth);
+		location[2].v = (v + vHeight) / (float)textureWidth + (.1f / (float)textureWidth);
 		;
-		location[3].v =
-			(v + vHeight) / (float)textureWidth + (.1f / (float)textureWidth);
+		location[3].v = (v + vHeight) / (float)textureWidth + (.1f / (float)textureWidth);
 		;
 	}
 }
@@ -316,71 +301,61 @@ void StaticInfo::setNewUVs(float uLeft, float vTop, float uRight, float vBottom)
 	location[1].v = location[2].v = vBottom;
 }
 
-void drawShadowText(int32_t colorTop, int32_t colorShadow, HGOSFONT3D font,
-	int32_t left, int32_t top, bool proportional, PCSTR text, bool bold,
-	float scale)
+void drawShadowText(int32_t colorTop, int32_t colorShadow, HGOSFONT3D font, int32_t left,
+	int32_t top, bool proportional, PCSTR text, bool bold, float scale)
 {
-	drawShadowText(colorTop, colorShadow, font, left, top, proportional, text,
-		bold, scale, -1, 1);
+	drawShadowText(colorTop, colorShadow, font, left, top, proportional, text, bold, scale, -1, 1);
 }
 
-void drawShadowText(int32_t colorTop, int32_t colorShadow, HGOSFONT3D font,
-	int32_t left, int32_t top, bool proportional, PCSTR text, bool bold,
-	float scale, int32_t xOffset, int32_t yOffset)
+void drawShadowText(int32_t colorTop, int32_t colorShadow, HGOSFONT3D font, int32_t left,
+	int32_t top, bool proportional, PCSTR text, bool bold, float scale, int32_t xOffset,
+	int32_t yOffset)
 {
-	gos_TextSetAttributes(
-		font, colorShadow, scale, false, proportional, bold, false, 0);
+	gos_TextSetAttributes(font, colorShadow, scale, false, proportional, bold, false, 0);
 	gos_TextSetRegion(0, 0, Environment.screenWidth, Environment.screenHeight);
 	gos_TextSetPosition(left + xOffset, top + yOffset);
 	gos_TextDraw(text);
-	gos_TextSetAttributes(
-		font, colorTop, scale, false, proportional, bold, false, 0);
+	gos_TextSetAttributes(font, colorTop, scale, false, proportional, bold, false, 0);
 	gos_TextSetPosition(left, top);
 	gos_TextDraw(text);
 }
 
-void drawShadowText(int32_t colorTop, int32_t colorShadow, HGOSFONT3D font,
-	int32_t left, int32_t top, int32_t right, int32_t bottom, bool proportional,
-	PCSTR text, bool bold, float scale, int32_t xOffset, int32_t yOffset)
+void drawShadowText(int32_t colorTop, int32_t colorShadow, HGOSFONT3D font, int32_t left,
+	int32_t top, int32_t right, int32_t bottom, bool proportional, PCSTR text, bool bold,
+	float scale, int32_t xOffset, int32_t yOffset)
 {
-	gos_TextSetAttributes(
-		font, colorShadow, scale, true, proportional, bold, false, 2);
-	gos_TextSetRegion(
-		left + xOffset, top + yOffset, right + xOffset, bottom + yOffset);
+	gos_TextSetAttributes(font, colorShadow, scale, true, proportional, bold, false, 2);
+	gos_TextSetRegion(left + xOffset, top + yOffset, right + xOffset, bottom + yOffset);
 	gos_TextSetPosition(left + xOffset, top + yOffset);
 	gos_TextDraw(text);
-	gos_TextSetAttributes(
-		font, colorTop, scale, true, proportional, bold, false, 2);
+	gos_TextSetAttributes(font, colorTop, scale, true, proportional, bold, false, 2);
 	gos_TextSetRegion(left, top, right, bottom);
 	gos_TextSetPosition(left, top);
 	gos_TextDraw(text);
 }
 
-int32_t interpolateColor(int32_t color1, int32_t color2, float percent)
+uint32_t interpolateColor(uint32_t color1, uint32_t color2, float percent)
 {
-	int32_t color = 0xffffffff;
+	uint32_t color = 0xffffffff;
 	if (percent > 1.0)
 		color = color2;
 	else if (percent < 0.0)
 		color = color1;
 	else
 	{
-		int32_t alphaMin = (color1 >> 24) & 0xff;
-		int32_t alphaMax = (color2 >> 24) & 0xff;
-		int32_t newAlpha =
-			(float)alphaMin + ((float)(alphaMax - alphaMin)) * percent;
-		int32_t redMin   = (color1 & 0x00ff0000) >> 16;
-		int32_t redMax   = (color2 & 0x00ff0000) >> 16;
-		int32_t newRed   = (float)redMin + ((float)(redMax - redMin)) * percent;
-		int32_t greenMin = (color1 & 0x0000ff00) >> 8;
-		int32_t greenMax = (color2 & 0x0000ff00) >> 8;
-		int32_t newGreen =
-			(float)greenMin + ((float)(greenMax - greenMin)) * percent;
-		int32_t blueMin = color1 & 0x000000ff;
-		int32_t blueMax = color2 & 0x000000ff;
-		int32_t newBlue =
-			(float)blueMin + ((float)(blueMax - blueMin)) * percent;
-		color = newBlue + (newGreen << 8) + (newRed << 16) + (newAlpha << 24);
+		uint32_t alphaMin = (color1 >> 24) & 0xff;
+		uint32_t alphaMax = (color2 >> 24) & 0xff;
+		uint32_t newAlpha = alphaMin + ((alphaMax - alphaMin) * percent);
+		uint32_t redMin   = (color1 & 0x00ff0000) >> 16;
+		uint32_t redMax   = (color2 & 0x00ff0000) >> 16;
+		uint32_t newRed   = redMin + ((redMax - redMin) * percent);
+		uint32_t greenMin = (color1 & 0x0000ff00) >> 8;
+		uint32_t greenMax = (color2 & 0x0000ff00) >> 8;
+		uint32_t newGreen = greenMin + ((greenMax - greenMin) * percent);
+		uint32_t blueMin  = color1 & 0x000000ff;
+		uint32_t blueMax  = color2 & 0x000000ff;
+		uint32_t newBlue  = blueMin + ((blueMax - blueMin) * percent);
+		color			  = newBlue + (newGreen << 8) + (newRed << 16) + (newAlpha << 24);
 	}
 	return color;
 }
